% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_exec.R
\name{exec_stop}
\alias{exec_stop}
\title{Error while executing Vim command}
\usage{
exec_stop(message, command, suggestion)
}
\arguments{
\item{message}{statement that describes the error that occurred.}

\item{command}{the Vim command that was being attempted.}

\item{suggestion}{suggest a possible cause/fix for the error.}
}
\value{
Exits R execution.
}
\description{
Helper function for rsvim_exec
}
