% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_vim_commands.R
\docType{data}
\name{rstudio_vim_commands}
\alias{rstudio_vim_commands}
\title{RStudio Vim Commands data.frame}
\format{
A data frame with < 50 rows and 2 columns:
\describe{
\item{name}{Full command name}
\item{short_name}{Shorthand that can be used instead}
}
}
\source{
\url{https://github.com/rstudio/rstudio/blob/main/src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/keybinding-vim-uncompressed.js}
}
\usage{
rstudio_vim_commands
}
\description{
A data.frame of commands valid for use taken from the RStudio Vim key-binding
javascript file. See pre-processing in \code{create_rstudio_vim_commands.R}. This
is not a fully comprehensive list as it doesn't include commands like \code{help}
and \code{R}.
}
\keyword{datasets}
