% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_add_command.R
\name{rsvim_add_command}
\alias{rsvim_add_command}
\title{Append command to config file}
\usage{
rsvim_add_command(command, comment = "", to = rsvim_default_path())
}
\arguments{
\item{command}{String. A Vim ex command.}

\item{comment}{String. Text describing the purpose of the command. This will
be written one line above the command, preceded by double-quotes \verb{"} to be
treated as a comment.}

\item{to}{String. File path to config file.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Append a Vim ex command to the end of a config file, optionally preceded by a
descriptive comment.
}
\examples{
\dontrun{
rsvim_add_command("imap jk <Esc>")

rsvim_add_command("map \o o<Esc>",
                  "create a new line and stay in normal mode")
}
}
