% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_default_file.R
\name{rsvim_default_file}
\alias{rsvim_default_file}
\title{Get suggested config file path}
\usage{
rsvim_default_file()
}
\value{
String. The full file path to \code{.vimrc}
}
\description{
The file is \code{.vimrc} located in
\verb{\%APPDATA\%/RStudio/keybindings/} where you can also find RStudio's JSON
keybinding config files.
}
\examples{
\dontrun{
file.edit(rsvim_default_file())
}
}
\seealso{
\code{\link[=rsvim_exec_file]{rsvim_exec_file()}} To execute commands from a file at this path.
}
