% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_is_valid_command.R
\name{rsvim_is_valid_command}
\alias{rsvim_is_valid_command}
\title{Check if a Vim ex command is valid}
\usage{
rsvim_is_valid_command(command)
}
\arguments{
\item{command}{String. Command to check for validity.}
}
\value{
Logical. Whether the command is supported in RStudio's javascript
implementation of Vim.
}
\description{
Checks against the RStudio defaultExCommandMap in
\href{https://github.com/rstudio/rstudio/blob/main/src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ace/keybinding-vim-uncompressed.js}{keybinding-vim-uncompressed.js}.
The list is not complete and doesn't include for example \verb{:help} and \verb{:R}.
}
\examples{
# valid commands
rsvim_is_valid_command("map")
rsvim_is_valid_command("noremap")

# not valid commands
rsvim_is_valid_command("hello")

# it is not a complete list at present:
rsvim_is_valid_command("help")
}
\seealso{
\link{rstudio_vim_commands} for all valid commands.
}
