% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_use_template.R
\name{rsvim_use_template}
\alias{rsvim_use_template}
\title{Build a Vim config from a template}
\usage{
rsvim_use_template(
  to = rsvim_default_path(),
  from = rsvim_example_path(),
  overwrite = FALSE
)
}
\arguments{
\item{to}{String. File path to write the Vim config file.}

\item{from}{String. File path to the config file to copy from.}

\item{overwrite}{Logical. By default this function will error if there is
already a file at the location. Use \code{overwrite = TRUE} to replace the existing
file at \code{to} with the contents of \code{from}.}
}
\value{
Returns the \code{to} path invisibly.
}
\description{
Create a new config file from an existing config file. By default this is
based on the
\href{https://github.com/davidfoord1/rstudiovim/blob/main/inst/basic.vimrc}{basic.virmc}.
}
\examples{
\dontrun{
rsvim_use_template()

rsvim_use_template(
  "path/to/my/custom/.virmc",
  "path/to/template/.virmc"
)
}
}
