% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvim_exec.R
\name{rsvim_exec}
\alias{rsvim_exec}
\title{Execute vim command in source editor}
\usage{
rsvim_exec(command, focus_source = "Ctrl+1")
}
\arguments{
\item{command}{String. A vim command to execute. This is the text you would
type in the vim command dialogue box after \code{:} and before \code{Enter}.}

\item{focus_source}{String. Keyboard shortcut for focusing on the source
pane, e.g. as in Tools --> Modify Keyboard Shortcuts.... The RStudio
default is \code{Ctrl+1}. Build strings as appropriate for
\code{\link[KeyboardSimulator:keybd.press]{KeyboardSimulator::keybd.press()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Pass a string to be executed as vim command. \code{\link{KeyboardSimulator}} is used
to simulate navigating to source, and typing
}
\examples{
# Create a home-row binding jk to go from insert mode to normal mode.
rsvim_exec("imap jk <Esc>")

# Replace "text" with "rep" everywhere in the active document.
rsvim_exec("\%s/text/rep/g")


}
